youtube-trending/requirements.txt
a2wsgi==1.10.8
agate==1.9.1
aiologic==0.14.0
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==3.0.0
apache-airflow-core==3.0.0
apache-airflow-providers-common-compat==1.6.1
apache-airflow-providers-common-io==1.5.4
apache-airflow-providers-common-sql==1.26.0
apache-airflow-providers-smtp==2.0.3
apache-airflow-providers-standard==1.1.0
apache-airflow-task-sdk==1.0.0
argcomplete==3.6.2
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
backports.strenum==1.3.1
blinker==1.9.0
cadwyn==5.3.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.2
daff==1.3.46
dbt-adapters==1.14.7
dbt-common==1.23.0
dbt-core==1.9.4
dbt-extractor==0.6.0
dbt-postgres==1.9.0
dbt-semantic-interfaces==0.7.4
deepdiff==7.0.1
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.1.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
leather==0.4.0
libcst==1.7.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.14
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
msgpack==1.1.0
msgspec==0.19.0
networkx==3.2.1
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
ordered-set==4.1.0
packaging==25.0
parsedatetime==2.6
pathspec==0.12.1
pendulum==3.1.0
pluggy==1.5.0
protobuf==5.29.4
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
retryhttp==1.3.2
rich==14.0.0
rich-argparse==1.7.0
rich-toolkit==0.14.3
rpds-py==0.24.0
setproctitle==1.3.5
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.46.2
structlog==25.3.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.0.1
text-unidecode==1.3
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.4.0
uuid6==2024.7.10
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wirerope==1.0.0
wrapt==1.17.2
zipp==3.21.0
